pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "opentelemetry-exporter-otlp-proto-http>=1.40.0",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.61b0",
    "opentelemetry-sdk>=1.40.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
    "pyright>=1.1.390",
    "alembic>=1.18.4",
]

[tool.ruff]

target-version = "py313"
line-length = 100
indent-width = 4

exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "venv",
    "build",
    "dist",
]

lint.select = [
    "F",
    "B",
    "SIM",
    "C4",
    "UP",
    "TCH",
    "ASYNC",
    "FAST",
    "S",
    "N",
    "PT",
    "DTZ",
    "PERF",
    "I",
    "PL",
    "ARG",
    "ANN",
]

lint.ignore = [
    "E",
    "W",
    "B008",
    "S104",
    "S105",
    "S106",
    "S107",
    "S603",
    "S607",
    "N803",
    "N806",
    "PLR0913",
    "PLR0911",
    "PLR0912",
    "PLR2004",
    "PT011",
    "TC001",
    "TC002",
    "TC003",
]

[tool.ruff.lint.flake8-bandit]

check-typed-exception = true

[tool.ruff.lint.isort]

combine-as-imports = true
split-on-trailing-comma = true
known-first-party = ["app"]

[tool.ruff.lint.flake8-type-checking]

strict = true

[tool.ruff.lint.per-file-ignores]

"__init__.py" = ["F401"]
"migrations/versions/**" = ["PLR0915", "ANN"]
"tests/**" = [
    "S101",
    "S105",
    "S106",
    "PLR2004",
]

[tool.ruff.format]

quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.pyright]

pythonVersion = "3.14"
typeCheckingMode = "strict"

include = ["app"]
exclude = [
    "**/__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
]

reportMissingTypeStubs = true

reportUntypedBaseClass = false
reportUntypedFunctionDecorator = false

reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false

reportMissingImports = true
reportMissingModuleSource = false

reportIncompatibleVariableOverride = false

reportImplicitOverride = true
reportIncompatibleMethodOverride = true

useLibraryCodeForTypes = true
